Rename SafeUploadSessionDelegate to StreamableUploadSessionDelegate
Split into separate files and use a more descriptive name that
communicates the class makes all upload types streamable.
