+++ Modules/Sources/NetworkingCore/Network/UploadStreamProvider.swift
@@ -2,7 +2,7 @@ import Alamofire
 import Foundation
 
 /// Tracks `URLSessionTask` → `UploadRequest.Uploadable` mappings so that
-/// `SafeUploadSessionDelegate` can provide the correct `InputStream` when
+/// `StreamableUploadSessionDelegate` can provide the correct `InputStream` when
 /// URLSession calls `needNewBodyStream` for non-stream uploads.
 ///
 /// Alamofire's default `SessionDelegate` crashes with `fatalError` when
@@ -70,36 +70,5 @@ final class UploadStreamProvider: EventMonitor, @unchecked Sendable {
             }
         }
         lock.unlock()
-    }
-}
-
-/// A `SessionDelegate` subclass that safely handles `needNewBodyStream` for
-/// `.data` and `.file` upload types, preventing the crash in Alamofire's
-/// default implementation which only supports `.stream`.
-///
-/// For `.stream` uploads and untracked tasks, the superclass implementation
-/// is used unchanged.
-///
-/// See: https://a8c.sentry.io/issues/WOOCOMMERCE-IOS-1PHM
-///
-final class SafeUploadSessionDelegate: SessionDelegate, @unchecked Sendable {
-
-    let uploadStreamProvider = UploadStreamProvider()
-
-    override func urlSession(
-        _ session: URLSession,
-        task: URLSessionTask,
-        needNewBodyStream completionHandler: @escaping (InputStream?) -> Void
-    ) {
-        switch uploadStreamProvider.uploadable(for: task.taskIdentifier) {
-        case .data(let data):
-            completionHandler(InputStream(data: data))
-        case .file(let url, _):
-            completionHandler(InputStream(url: url))
-        case .stream:
-            super.urlSession(session, task: task, needNewBodyStream: completionHandler)
-        case .none:
-            completionHandler(nil)
-        }
     }
 }
