+++ Modules/Sources/NetworkingCore/Network/WordPressOrgNetwork.swift
@@ -173,7 +173,7 @@ private extension WordPressOrgNetwork {
         // The uploadStreamProvider event monitor tracks which upload data
         // belongs to each URLSessionTask so the delegate can provide it
         // when needNewBodyStream is called.
-        let delegate = SafeUploadSessionDelegate()
+        let delegate = StreamableUploadSessionDelegate()
         return Alamofire.Session(
             configuration: sessionConfiguration,
             delegate: delegate,
